small fixes, nagative training, dropout moved after wiq

+++ SQuAD/layers/TrainNAWLayer.py
@@ -29,7 +29,7 @@ class TrainNAWLayer(MergeLayer):
     def get_output_for(self, inputs, **kwargs):
         assert len(inputs) == 2
         input_, mask = inputs
-        naw_positions = mask.sum(axis=1) - 1
+        naw_positions = mask.astype('int32').sum(axis=1) - 1
         return TT.set_subtensor(input_[TT.arange(input_.shape[0]), naw_positions], self.W)
 
     def get_output_shape_for(self, input_shapes):
